// src/core/types.ts
export interface WPPost {
  id: number
  title: { rendered: string }
  slug: string
  link: string
  status?: string
  acf: {
    content: WPAcfElement[]
    favicon?: number
    logo?: number
  }
}

export interface WPAcfElement {
  acf_fc_layout: "element"
  type: "nosubs" | "subs"
  show_in_menu: boolean
  show_content_editor: boolean
  content_title: string
  content_main: string
  content_sub: WPAcfSubElement[] | null
}

export interface WPAcfSubElement {
  content_subtitle: string
  subcontent: string
}

export interface WPMedia {
  id: number
  source_url: string
}

export interface Destination {
  tabId: number
  elementIndex: number
  targetField: "content_main" | "subcontent"
  subIndex?: number
  sectionTitle?: string
  subSectionTitle?: string
}